First introduction of rotation grid
This patch adds the rotational grid, G_rot, to the MOM_control_struct.
For now, we will construct G and G_rot in unison, and then see whether
we truly need to keep both grids.

This only adds the variable, it does not actually manipulate or
otherwise rotate G_rot.  Will follow in later patches.