Pointers to dG and HI in MOM.F90
It's becoming clear to me that I need to build the grids, fields, etc on
the original grids, not the rotated ones.  So I am going to try and keep
dG and HI on the original grids for initialization.

Trying to use pointers so that we can just poing to the inputs when no
transformations are required (rotation, reflection, etc).