Input/rotated HI and dG; rotation functions
We now track both the input and rotated forms of HI (horizonal indices)
and dG (dynamic horizontal grid), so that we can use the unmodified
functions for generating grids and other input data.

Function for rotating the grid fields have also been implemented, but
are not yet used anywhere.

The total changes are not implemented, and rotated runs now crash since
the values of dG are garbage.