Metric implementation, asymmetric invariance
Metric grids can now be rotated, and the 90° rotation checksums agree
with the nonrotated configuration.  The new rotate_dyngrid function is
largely doing its job.

There is a larger issue with symmetric grids, where a symmetric field
(e.g. dxCu) will be rotated from (0:M,1:N) to (1:N,0:M), but will still
expect many loops over is[cdg]B to ie[cdg]B, which seems to be causing
problems, such as out-of-bound indexing in data domains.

I am not currently sure how to solve this, and it may not even be
solvable.