Metric rotation
We correct the mappings of the various fields, so that u points go to v
points, and dx's go to dy's (and vice versa).

Some might not be right, like maybe grad(f) needs a sign change?   Also
only did the 90° turn.

Also added output to ocean_geometry for the rotated grid, to check
output.  Looks OK!