Return G_in as grid state + pointer rework
We want the forcing setup functions to use the original grid, so we want
MOM to return the nonrotated grid when queried by
get_MOM_state_elements.

This required some overall reworking of the interior pointers, e.g. CS%G
is now a pointer rather than a target, and we have to formally construct
the full G and G_in grids, rather than just the active grid (G).

This is now more consistent, although we're now failing (as expected)
when we try to use the forcing fields like taux.  But this is expected
since the grids no longer match (by design!).

Next patch will add a rotation function for the fields.