Mech forcing rotation
The main change here is that we now rotate the mechanical coupler
forcing terms (wind stress, pressure, etc), which corrects the checksums
of those fields.

We have not yet rotated the non-mechanical forcings (heat, precip, etc)
so answers still do not agree, although there is now a greater number of
invariant checksums.

This change will break most diagnostic operations by the coupler,
especially with respect to forcings, since it still presents fields as
if part of the input grid, rather than the rotated grid seen by the diag
mediator.  I'm not yet sure how to resolve this.

This code will crash, although it works in my tests because I've
disabled most diagnostic output directed by the coupler (in uncommitted
code).