Rotation of input fluxes
Fluxes are now fully rotated; this works, although diagnostics are still
broken.  There are also funny issues with a flag inside fluxes, since it
seems to be used to track things and is also modified internally by MOM;
this is a problem for us because we copy the original fluxes struct and
miss these changes.

This runs but answers still differ;  This seems to be due to the
surface state variable, which also need to rotate..