Surface state rotation; flux flag fix
Mostly this rotates the surface state, so that the coupler sees the
original grid but MOM manipulates its mirror in the model.

This is early days, probably not working.  But it builds.

The use_fluxes flag is also mirrored across the fluxes struct, which
resolved some of the safety checks in the coupler.