Permutation of h-point shifted checksums
The checksums of shifted h-points for corners is now implemented, and a
rotation keeps these checksums invariant.

I have also fixed a bug in uvchksum with respect the sign of the
uv-field.  There may still be an error here, since we are effectively
assuming that these are vectors, and I'm not sure this is correctly
handled for scalars, which should not change sign.