Checksum scalar/vector support; 3d fixes
This patch modifies the uv checksum functions to distinguish between
scalar and vector arguments, the latter of which require a sign change
for specific rotations.

This patch also extends the work to generalize domain shift checksums of
uv fields from 2d to 3d.

Finally, the new scalar flags are added for a few sample fields.