Rotated checksum fixes
This is a partial fix to the checksums in the rotated domains.  Two
fixes here:

1. We shift the compute grid checksum (bc0) to always align with the
original (unrotated) compute grid.

2. We compute the default (west/south) shifted checksum so that it
matches the shift of the original unrotated grid.

Despite the potential for answer changes, this appears to have fixed
many observed inconsistencies and debug output is now in greater
agreement.

This is not complete however, many shifted checksums (corner, NSEW)
still have not been corrected.