fluxes%heat_added rotate; de-rotate to fluxes_in
This patch incorporates the heat_added field of fluxes into the general
flux rotation.

We also now copy a de-rotated fluxes back onto fluxes_in, while removing
some of the ad-hoc flag syncing which should no longer be needed.  This
is because fluxes can actually be changed by the model within a step.
While it doesn't seem necessary here, it is probably more correct to
make this change.

This patch is the final fix for the TC2 test.  After fixing a detected
bug (work_u in MOM_thickness_diffuse, ~L1236), the TC2 test now passes
the 90-degree rotation test.