Phased out the "parity domain"
The parity domain was an attempt to define the original compute domain
of a rotated grid for use in the checksums.  This is no longer needed,
since we now compute offsets inside of the checksum functions.

The parity domain was deemed unnecessary since one already has to
account for so many other factors that it's not useful to cache one
additional offset calculation.