Initial state rotation
This patch constructs the initial state on the input grid (G_in) and
rotates the u, v, h, T, S fields onto the model grid G.

We now see rotational reproducibility of the tc1 runs.

It also fixes a possible bug with sfc_state, which was being referenced
before it was created.  This was probably meant to be sfc_state_in.
Fixing this resolved some bugs in tc0 and tc3, but not all of them.