Non-thermodynamic init rotation
This fixes bugs with the initial state rotation when T and S are not
defined (non-thermodynamic mode).  The patch is messy but it does fix
the tc0 test case

(Up to a negative-zero in the CFL criterion, that is; will into it
later)

TC3 proceeds past its error as well but a new error is appearing.