Array rotation interface; start on OBC rotation
The rotate_* functions have been moved under a new module, with
interfaces for simplified access.

All of the targeted rotation functions are still explicitly calling
rotate_quarter, but are now doing so through an interface which can
potentially support integers.

Also started on the OBC rotation, which will unfortunately be a large
task.  Currently the "rotation" only copies (or points to) the original
content.

Also noting here that we're doing a lot of pointing to anonymous data
which was created by allocation of stack pointers.  I am consulting the
community to see if this violates anything in Standard Fortran.