segment inout fix; diagnostic stuff
The most important patch here is that an OBC segment was incorrectly
tagged as intent(out) in a function, which was clobbering its values.
It's now correctly tagged as intent(out) which fixed lots of problems.

Most of the patch is diagnostics garbage, which I need to discuss with
Alistair tomorrow morning.