Force mem cleanup; safer initialization
This patch attempts to start fixing some of the memory allocation
problems which are appearing in TC4.  It does not solve the underlying
problem, but is a safer code.

The major change is that we now deallocate forces and fluxes, the
rotated forms of input forcings, and the end of each timestep.  We thus
re-allocate and deallocate these on every timestep (step_MOM call).
This is far from efficient, but it's better than the weird quasi-static
variable assumption we were relying on previously.

We also now copy the thin_halo_updates field from G%Domain, which was
not being transferred in the clone_MD_to_MD function (G_in copied to G).

Finally, we now initialize dG_in to NULL().

Only the first change here would have significant impact, but the amount
of bad memory accesses reported by valgrind has reduced...