Rotate surface flux diagnostics
Same deal as mech forcing; we rotate the fluxes back to the model grid
when computing and writing diagnostics.

Slightly trickier because additional diagnostics are computed in a
scratch array, so its size was changed from G to diag%G.  I am probably
not even using this macro, so no idea if this is even being respected.

surface state is not checked, and is probably broken.