Restart state rotation
This patch adds support for rotation of restarts, specifically
introducing the following concepts:
    - u-v pairs are swapped during registration to preserve field names
      and grid sizes.
      NOTE: This is currently manual, but I would like to introduce a
      "register_restart_uvfield" function to absorb a lot of this
      boilerplate code
    - field_[234]d pointers were introduced to save_restart, which
      rotate the fields just before passing to `write_field`.
      NOTE: This is still incomplete since we do not account for
      sign switches of vectors.
    - a rotate_array_real_4d function was added to the interface to
      support 4d restart fields (still assuming (x,y,-,-)

This is not perfect, but does preserve array sizes and (hopefully)
values, and no longer requires any corrections to the driver code.

Finally, the internal t=0 restart, GOLD_IC.nc uses the input grid,
rather than the model grid, and should now also match the non-rotated
state.

The driver is now oblivious to the internal rotation when producing
restart files!