OBC restart rotation; restart bugfix; B-chksums
First change is that two of the OBC restart field pairs are now swapped.
The third pair is now yet handled correctly, since the allocation and
assignment are a bit too tangled to sort out quickly.  (TODO?)

Second change is a minor fix to `turns` in one of the restart
functions, which was not updated correctly.

Last change is the first fix to the B-point checksums, so that the stats
are now correctly reported on the original compute domain.  The
bitcounts and the halo-shifted bitcounts are still not working
correctly.