MOM ALE sponge rotation (pt1: index)
This first patch rotates the columnar index information of the ALE
sponge.  it does so by re-constructing the Iresttime and (optionally)
data_h arrays, rotating them, and then using initialize_ALE_sponge_*
functions to re-construct the rotated ALE sponge.

The fields are not yet handled (set_ALE_field-like function) but will
address this in a later patch.