time_interp_external update in horizontal remap
This patch rotates the output from time_interp_external calls used in
the horizontal remapping function based on FMS ID.  This is not a great
solution, but is mostly to demonstrate how we might proceed.

This does not fix the issues with sponges, since the internal pointer to
the sponge tracer field is incorrect.  A comment was added to explain
this problem.

Finally the non-ALE sponge rotation was removed, since it does not yet
exist (oops).