ALE sponge rotation fixes
Fixed some bugs in the ALE sponge rotation, and also added a function to
fix up the variable pointers.

By all accounts this appears to be working, and the ALE sponge is
successfully rotated.

The ocean.stats output is still different, however; it seems to be
related to the depth list.  Still looking into it.