Hypothetical setup for configurable CVMix
This patch proposes a potential method for maintaining CVMix as an
independent library rather than as symbolic links in the MOM6 source
code.

Here, CVMix is handled inside the `ac/deps` directory in the same manner
as FMS.

This would also put a larger burder on developers to maintain their own
CVMix builds, rather than pushing it into the MOM6 build.  But it would
also offer greater flexibility to reconfigure CVMix.

This does not address all the potential issues with this approach, but
it offers proposed solutions to the various issues that could arise.