Profile generation in .testing; GHAction support
The .testing Makefile was updated to parse the profile run stdout
content to produce the clocks.json file.  Also the file was renamed from
prof.json to clocks.json.

The clock parsing tool was updated to not pretty-format the output on
default, and now uses a flag to control this behavior.  Might change the
default in the future.

An experimental GitHub Actions config was added to run the profile
rules.  I have never done this before, so it may not actually do
anything, but let's see what happens.