Integrate perf into performance testing
This adds a `perf` test to the Makefile and a (probably futile) attempt
to run it on the GH Actions cloud.

I've also bundled a script to read and parse the perf.data output and
return an equivalent JSON output file.

Not expecting a whole lot here but let's see what happens...