(*)non-Boussinesq Z unit update_OBC_segment_data
  Revised the update_OBC_segment_data code to keep the z-space input data in
height units ([Z ~> m]) rather than rescaling them to thickness units.  This
change means that the non-Boussinesq open boundary condition calculations avoid
using the Boussinesq reference density.  Associated with this change, 5 internal
variables in update_OBC_segment_data were renamed to reflect that they are layer
vertical extents instead of thicknesses.  Also added or amended comments
describing the purpose and units of 25 real variables in this module and
corrected a handful of lines in this module that were not adhering to the
guidance in the MOM6 style guide.  Answers are bitwise identical in any
Boussinesq cases.  However, answers will change in any non-Boussinesq cases that
use open boundary conditions that use time_interp_extrnal to read in Z-space
data.