Perf: Better error report for failed parsing
Print the line and its tokenization in parse_perf.py if the tokenization
fails for any reason.  This is most likely to occur if the fourth token
is not an integer.

This is possibly a platform-dependent issue, and not something easily
replicated on any machine we are using, so it is easier to simply
integrate it into the codebase.