2D and 3D methods for EOS density and spec volume
This patch adds two methods to support 2D and 3D arrays in
calculate_density and calculate_spec_vol.  Only arrays without halos are
supported, so as to support optimized contiguous loops.  The interface
for these methods excludes any "domain" type containing index bounds.

Only the generic base type is currently supported, meaning that any such
calculations will be replaced with loops of calls to the elemental
function, and will not be optimized.  This would be resolved by
including equivalent functions in each EOS type, which I would expect to
happen in a later commit.