Refactor density integrals to use 2D and 3D arrays
This patch replaces the packed 1D arrays in the various density
integrals with equivalent 2D and 3D arrays.  This eliminates the need to
introduce "position" index calculations within the packed arrays.

As long as the arrays are contiguous, as assumed by their simple
structure, the performance should be equivalent to the packed 1D arrays.