Autoconf: Disable LDFLAGS for C testing
Certain compilers would expect values passed to ld which cannot be
passed to the C compiler.  This caused the autoconf C configuration to
fail, since it uses to CC for linking rather than LD.

Since we never use the C compiler for linking, this is not something we
need to be concerned about, and the LDFLAGS is disabled when testing the
C compiler.

This is only a concern when using certain GPU-related flags in the
Nvidia compiler, and is only a problem for libraries with a significant
amount of C code, such as FMS.