horvisc k-loop with present(G)
It seemed like there was some implicit `G` fields being copied without
our knowledge.  If `present(G)` is absent from acc kernels, then it just
implicitly copies everything that is missing.

This can only cause errors when the parent derived type is not copied.
What is probably happeneing:

* `G%IdyCu` is referenced in the loop, so it is implicitly copied.
* `G` is *NOT* in the loop, so it is *NOT* copied.

So you have the contents of `G` but not `G` itself, which presumably
causes errors.  (Not sure which, come back to this.)

We do not want to do `G` field copies on every call, so we want to disable
these implicit copies.  We can do this with the `present(G)` attribute.

Although we were copying `G` and its contents, we were not doing it at
hte correct point.  `G` had been established, (thus averting a "partial
presence" error) but the contents of `G` were all wrong, particularly
the addresses (thus "Illegal address" errors).

Once the copies were moved to the correct position, the errors were
removed and the runs appear to be working correctly once again.

My personal thoughts: This is NOT FUN to work out.  We are getting
almost no information from the compiler about these problems.  Should
consider looking at some other openacc compilers (even those which don't
support GPUs).