autoconf: Potential autoheader integration
Mostly just playing with this for now, but there are a few gaps.

Mainly, it's not clear to me how to prevent C-like comments in config.h.

In practice, it's not an issue because only preprocessed (*.F90) files
would include these files, and strictly these are permitted in the CPP
preprocessor - even in Fortran.  But I still don't think it's a good
idea.

Anyway, this is a work in progress, and should be taken as such.