MEKE: Remove redundant if-block in Strang split
In the second stage of Strang splitting, there is a check for `sdt >
sdt_damping`, following a check for nonzero `KH` or `K4`.

However, `sdt_damp` can only be less than `sdt` when these are nonzero.
Otherwise, `sdt_damping` equals `sdt`.

Best as I can tell, this check is unnecessary and potentially
problematic for optimization.  This patch removes the redundant check.