Migrate main PF[uv] loop to GPU
This is a *very* minor first draft to migrate the main pressure force
loop to the GPU device.

The reduced gravity Montgomery potential dM loop is also migrated,
primarily to demonstrate that we can keep PF[uv] on the remote device
without copying it back over.  But it completely ignores the
intermediate calculation of dM (because it wasn't working.)

Getting there...