PF: Move Z_0p to GPU (barely)
I have moved the Z_0p compute loop to the GPU.  Not in any practical
sense, since it copy it right back to the CPU.

But there is some weird shit going on here: This only worked if I moved
the target region inside of the if(use_EOS) block.

When outside, I got the same NaN issues with BT PFu that I saw with
dz_geo.  As with those errors, it doesn't seem to actually depend on the
arrays, but on the presence and number of target regions, or some other
resource.

This is concerning and needs more observation.  We don't know why this
is failing.