Coriolis: First steps
Area_[hq] arrays are migrated to the GPU.  Pulled back to CPU for OBC
calculations.

The massive legacy OpenMP direction is blocked out for now, if only
because it completely breaks during GPU compilation.  (OBC causes a
bizarre error with a bunch of LLVM code.)

Placeholder !$omp declare target for gradKE.  Not sure if this will be
possible, still needs investigation.