MOM_murmur_hash: MurmurHash3 implementation
This patch adds the MOM_murmur_hash module, a non-cryptographic hash
function used to generate unique hashes for arrays.

The particular property of interest is that murmur hashes are sensitive
to order.  This will allow us to detect a wider range of answer changes
which would otherwise be undetected by the current min/max/mean/bitcount
method.