GitHub: FMS-MOM6 build jobs, tarball artifacts
Inconsistencies in jobs within a workflow were leading to differences in
compiler versions across jobs.  This was breaking GitHub Actions
workflows.  We want to continue using `*-latest` runners, so we have
merged the FMS stage into each of the MOM6 builds.

There are additional changes which improve the overall workflow/job
decomposition.

* FMS builds included in MOM6 stage

* Artifacts are now packed into tarballs, which preserve most of the
  file state (mtime, file mode, etc).

  * No longer any need to chmod the executables.

  * `*.gnco` timestap warnings are removed.

  * Overall simplification of generated executable lists; we now
    assume that build/* only contains executables.

* build.timing and build.unit are now correctly tagged as .NOTPARALLEL