Barotropic: Move CS alloc inside of init
The `!$omp target update to(CS)` inside of `barotropic_init` seemed
choke on some Nvidia compiler versions.  It was perhaps unnecessary,
since we could replace it will a well-timed `map(to: CS)` inside of
`barotropic_init`.

Naively swapping did not work because of a missing use of CS%frhat[uv]
happening on CPU-side.  (Perhaps this should not be, but for now it is
the case...)

By making both change, the `benchmark` case is again working.