DO_LOCALITY compatibility macro
This patch introduces a new macro to conditionally apply the locality
specifiers to do concurrent loops.

Currently there are only two instances: local() and reduce(), both in
MOM_barotropic.F90.  Both were introduced after do concurrent, local
with 2018 and reduce with 2023.  There are compilers in active use that
still do not support them.

They can now be wrapped with the DO_LOCALITY macro, which removes them
if the compiler does not support these features.

The macro is defined with the HAVE_FC_DO_CONCURRENT_LOCAL macro, which
can be set with either -D or with autoconf.  An M4 macro has been added,
invoked by configure.ac, which tests for local().

Someday it might be a good idea to create separate tests for each
locality spec.

A minor patch was also required to makedep to handle #define statements
without values, although macro functions are still not properly handled.