Barotropic: Stricter GPU copies for halo updates
GPU-CPU copies are only triggered if the domain is reentrant in x or y.
Otherwise, halo updates would not have any impact.

This is still fairly limited in scope.  It assumes one MPI rank, and
doesn't account for tripoles, which have their own flag.

Some local variables are also declared, which removed a few transfers.