More consistent CS alloc/upload setup
This new pattern does CS allocations before the initialization which
fills their values.

The CS then uploads after all the various scalar values have been set,
but before the arrays have been populated.

Allocations are *outside* the init function, and uploads are *inside*
the function.

This avoids some of the confusing patterns have have existed, where we
sometimes set up a CS before or after an init function.

Note that this is largely working around the confusing (if not wrong)
pattern of Nvidia.  An `update to (CS)` will wipe out an existing array
(at least I think so), which is what causes much of the drama about
setup.  As long as we separate scalar and allocatable setup, then we
should be good.