Continuity CS outside of init
This moves the Continuity CS to the dycore init function.  For some
reason, this avoids an answer change with CPU.  (Possibly because alloc
inside of a function doesn't quite match the CS outside of it?)

A few minor data transfers are also added to fix up differences in the
chksum log output.