Continuity: Add locality to do concurrent
Do concurrent inside of !$omp target teams loop seems to fail standard
openmp tests if locality is not correctly set.

This patch adds the correct locality to two of the four loops, and
(apparently) part of what's required for the other two.

I ran over 100 tests, and the two completed loops passed.  The other two
were commented out because they still do not pass.

I have a hunch that domore is the culprit, but I've done everything to
account for and rule it out, and yet they still fail.  Still more work
is needed, it seems.