Vertvisc: formatting/style cleanup
This patch reduces the formatting changes to this branch and brings it
closer to both dev/gfdl and the MOM6 style guide.

One minor non-cosmetic change is the reversion of ADp%dv_dt_str(i,J,1)
calculation into one of the main Thomas loop.  This returns it to a
separate loop.