Vert friction: Explicit CS alloc in dycore init
This patch removes the NVIDIA compiler preprocessing and explicitly
compiles the `CS` in the dynamic core initilization.

`vertvisc_init` is no longer responsible for allocation.  The allocation
test has also been removed.  We just have to trust each other now (or
segfault when it refs an unallocated field).

The previous error in the CI was the absense of `CS` allocation in the
other timestep methods (unsplit, unsplit RK, split RK2b).