Force FMAs in vertvics thomas solvers
the b1 = 1. / (b1_denom + dt * a_[uv](k+1)) refuses to use FMAs in the
new solvers, since it caches a_[uv]*dt and uses it in the next
iteration.

I added a nasty if(dt < 0) exit statement to confuse the compiler and
ensure it won't try to rely on the next iteration.  (Although it really
shouldn't matter!  But not my problem)