autoconf: Build GibbsSeaWater (GSW) as deps
The GibbsSeaWater project has been typically inserted into MOM6 as
source code.  This leads to a few complications:

* It must be inserted as a submodule, with symlinks in the source code.

* The codebase is external, and cannot be expected to follow MOM6 style.

General management of the code requires unique manipulations to include
or exclude target directories and files.

This patch builds GSW as a separate model, following the patterns used
for FMS.  The autoconf build now creates a libgsw.a archive which is
used for any calls into GSW, rather than from compilations of symlinked
code.

To avoid disruptions to exisiting projects, the existing symlinks have
been preserved, as well as the pkg checkout, but these can be phased out
in the future.

TODO:
* GSW lacks the search checks of FMS, and actually relies on shared
  include and lib directories set by the FMS search.

* While we're pulling things apart, this is a chance to rename libFMS.a
  to libfms.a (following both the FMS team as well as most posix-like
  convnetions.)

* libFMS.a and libgsw.a follow almost identical patterns.  There is
  probably a simple way to merge them.