Autoconf: --enable-werror flag
Optional autoconf argument to enable -Werror during testing.

This flag is configured separately, since very few tests or even
dependencys can build without errors.  It is appended to the final
result.

Currently this assumes the GCC syntax, which is fairly common across
compilers, but we should also develop an an autoconf test to verify
and/or select the correct flag, if we choose to use this for other
compilers.